+++ cypress/integration/optionsPanel.spec.js
@@ -1,5 +1,6 @@
 /// <reference types="cypress" />
 
+// TODO - Stub these responses
 context("Options Panel", () => {
   beforeEach(() => {
     cy.fixture("jobs50").then((jobsJson) => {
@@ -30,5 +31,24 @@ context("Options Panel", () => {
     cy.get(
       "#app > div.search__container > div.options-panel__container > label:nth-child(1) > input[type=checkbox]"
     ).should("be.checked");
+  });
+
+  it("Should retain location search value", () => {
+    cy.get("#location-search").should("have.value", "");
+    cy.get("#location-search").type("Los Angeles");
+    cy.get(".search__button").click();
+    cy.wait(1000);
+    cy.get("#location-search").should("have.value", "Los Angeles");
+  });
+
+  it("Should retain options values", () => {
+    cy.get("#location-1").should("not.be.checked");
+    cy.get(
+      "#app > div.search__container > div.options-panel__container > label:nth-child(3) > span"
+    ).click();
+    cy.get("#location-1").should("be.checked");
+    cy.get(".search__button").click();
+    cy.wait(1000);
+    cy.get("#location-1").should("be.checked");
   });
 });