+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- OptionsPanel tests
+
 ### Changed
 
 - If "How To Apply" is a single link, display a styled link